:fire::art: removed unused imports and outscourced CodeChartsDataTypes to remove ktlint errors.
