+++ src/main/code/de/tuchemnitz/se/exercise/core/graphics/codecharts/CodeChartsDialogView.kt
@@ -19,11 +19,8 @@ import tornadofx.vbox
 /**
  * First view during use of CodeCharts. Allows user to start the tool.
  */
-class CodeChartsDialogView/*(
-    private val cssRule: CssRule = Style.ccDialogWrapper
-)*/ : View("Willkommen bei CodeCharts!") {
+class CodeChartsDialogView : View("Willkommen bei CodeCharts!") {
     override val root: BorderPane by fxml(MainApp.MAIN_VIEW_TEMPLATE_PATH)
-
     private val contentBox: VBox by fxid("content")
 
     init {
@@ -52,7 +49,7 @@ class CodeChartsDialogView/*(
         }
     }
 
-    // to be removed
+    // needed for git button
     fun printGitButton() {
     }
 }
